requirements.txt
# Oracle Database
cx_Oracle>=8.3.0

# Feishu SDK
lark-oapi>=1.2.0

# Configuration and utilities
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Rate limiting
ratelimit>=2.2.1
